backend/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
pydantic==2.11.7
python-dotenv==1.1.1
openai==1.102.0
langchain==0.3.27
langchain-openai==0.3.28
